decouple reader
