+++ src/lib.rs
@@ -17,6 +17,7 @@ pub mod modulators;
 mod node;
 mod processor;
 mod processors;
+mod reader;
 mod sources;
 
 pub use basic_types::*;
@@ -25,4 +26,5 @@ pub use manager::*;
 pub use node::*;
 pub use processor::*;
 pub use processors::*;
+pub use reader::*;
 pub use sources::*;
